// frontend/src/app/models/pregled.ts
export class Pregled{
    lekar: string = '';
    imeLekara: string = '';
    prezimeLekara: string = '';
    pacijent: string = '';
    naziv: string = '';
    trajanje: number = null;
    cena: number = null;
    datum: string = '';
    vreme: string = '';
    ogranak: string = '';
    objasnjenje: string = '';
}